Refactor tests
Move benchmarks into a separate file, with a feature flag "bench",
so that normal tests can compile on stable.
