+++ .github/workflows/rust.yml
@@ -19,4 +19,4 @@ jobs:
     - name: Build
       run: cargo build --verbose --locked
     - name: Run tests
-      run: cargo test "test" --verbose --locked
+      run: cargo test --verbose --locked
